pyproject.toml
[tool.poetry]
name = "desafio-python-django"
version = "0.1.0"
description = ""
authors = ["Marcus Gabriel <marcusgabriel.ds@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.1.3"
djangorestframework-simplejwt = "^4.5.0"
djangorestframework = "^3.12.2"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
django-test-without-migrations = "^0.6"
model_bakery = "^1.2.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
